Move url validation and email sender to own extensions folder, clean up config
