+++ src/utils/db_uri_builder.py
@@ -1,0 +1,12 @@
+def build_db_uri(
+    username: str | None, password: str | None, database: str | None, database_host: str
+) -> str | None:
+    if not all(
+        (
+            username,
+            password,
+            database,
+        )
+    ):
+        return None
+    return f"postgresql://{username}:{password}@{database_host}:5432/{database}"
