+++ src/utils/strings/url_validation_strs.py
@@ -19,6 +19,7 @@ ORIGINAL = "original"
 URL = "url"
 CONTENT_TYPE = "Content-Type"
 HTML_CONTENT = "text/html"
+URL_VALIDATION_MODULE = "urlValidationModule"
 
 
 class URL_VALIDATION:
@@ -43,3 +44,4 @@ class URL_VALIDATION:
     URL = URL
     CONTENT_TYPE = CONTENT_TYPE
     HTML_CONTENT = HTML_CONTENT
+    URL_VALIDATION_MODULE = URL_VALIDATION_MODULE
